+++ miden/tests/integration/air/chiplets/hasher.rs
@@ -52,6 +52,31 @@ fn mtree_set() {
 }
 
 #[test]
+fn mtree_verify() {
+    let asm_op = "mtree_verify";
+
+    let index = 3_usize;
+    let (leaves, store) = init_merkle_store(&[1, 2, 3, 4, 5, 6, 7, 8]);
+    let tree = MerkleTree::new(leaves.clone()).unwrap();
+
+    let stack_inputs = [
+        tree.root()[0].as_int(),
+        tree.root()[1].as_int(),
+        tree.root()[2].as_int(),
+        tree.root()[3].as_int(),
+        index as u64,
+        tree.depth() as u64,
+        leaves[index][0].as_int(),
+        leaves[index][1].as_int(),
+        leaves[index][2].as_int(),
+        leaves[index][3].as_int(),
+    ];
+
+    build_op_test!(asm_op, &stack_inputs, &[], store)
+        .prove_and_verify(stack_inputs.to_vec(), false);
+}
+
+#[test]
 fn mtree_merge() {
     let asm_op = "mtree_merge";